@@ -7,26 +7,27 @@
   "author": "Cody Ogden",
   "license": "MIT",
   "devDependencies": {
-    "@types/lodash": "^4.14.171",
-    "@types/react": "^17.0.16",
+    "@types/jest": "^27.0.1",
+    "@types/lodash": "^4.14.172",
+    "@types/react": "^17.0.19",
     "@types/react-select": "^4.0.17",
-    "@types/styled-components": "^5.1.12",
+    "@types/styled-components": "^5.1.13",
     "cz-conventional-changelog": "^3.3.0",
-    "jest": "27.0.6",
-    "jest-resolve": "27.0.6",
+    "jest": "27.1.0",
+    "jest-resolve": "27.1.0",
     "moment": "^2.29.1",
     "typescript": "^4.3.5"
   },
   "dependencies": {
-    "babel-plugin-styled-components": "^1.12.0",
+    "babel-plugin-styled-components": "^1.13.2",
     "date-fns": "2.23.0",
-    "next": "10.2.3",
+    "next": "^11",
     "prop-types": "^15.7.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-select": "4.3.1",
     "slugify": "1.6.0",
-    "styled-components": "5.3.0"
+    "styled-components": "5.3.1"
   },
   "scripts": {
     "dev": "next dev",
