@@ -1,7 +1,9 @@
 // Global Stylesheet
 import '../public/global.css';
 
-function MyApp({ Component, pageProps }) {
+import type { AppProps } from 'next/app'
+
+const MyApp: React.FC<AppProps> = ({ Component, pageProps }) => {
     return <Component {...pageProps} />
 }
 
